refactor(deck_builder_panel): split DeckBuilderPanel into mixin package
Move widgets/panels/deck_builder_panel.py into a package: frame.py
keeps the nested _SearchResultsView helper and the full _build_ui;
properties.py holds the translate helper and the filter/result
getters; handlers.py owns every _on_* callback, clear_filters,
update_results, radar integration, and the add-button refreshers.
