+++ widgets/panels/deck_builder_panel/__init__.py
@@ -1,0 +1,7 @@
+"""Deck builder panel package."""
+
+from __future__ import annotations
+
+from widgets.panels.deck_builder_panel.frame import DeckBuilderPanel
+
+__all__ = ["DeckBuilderPanel"]
